backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pandas
scikit-learn
google-play-scraper
app-store-scraper
playwright
langdetect
deep-translator
emoji
google-generativeai
transformers
numpy